Properly clean up when setting new InputFilter.
In f1b93471565c43e57e9e199596d51f7caaec8ebd, the single InputFilter that
is set in InputManagerService was replaced with a chain of filters that
are all listening for InputEvents.

The original field mInputFilter was replaced in this patch but not
removed. When a new InputFilter is added, we check that this unused
field mInputFilter != null before doing necessary teardown of the previous
filter. Since this is always null, this causes the previous filter to
not be disconnected when a new one is set with setInputFilter().

If the user toggles "Magnification Gestures" on and off twice in Accessibility
Settings, this will send the old and new InputFilters into a loop
sending and receiving touch events, locking up the device completely
until reboot.

Remove all references to the unused mInputFilter field.

Change-Id: I6ef67a664f1e783f3f0402b7c2a0984499a4b614
Signed-off-by: Eliminater74 <eliminater74@gmail.com>