Fix for reboot due to Volume State not updated in callback
When repeatedly formatting external sdcard as portable & internal
storage there is a possibility of state mismatch in App listener
volume state and actual volume state maintained by vold. This
results in reboot when try to create User directory in package-
manager when the volume is unmounted

The change verifies current volume state as well when volume state
change callback is received in package-manager. This avoids the
problem of volume state mismatch

Change-Id: I018a1a15dec72e86950293819d4e8481029b6dd2
Issue-Id: http://code.google.com/p/android/issues/detail?id=203886
Signed-off-by: Shasidhar Ganiga <sganig@codeaurora.org>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>