Introduce FP unlock boost
This patch adds a boost which is meant to speed up
unlocking using the fingerprint sensor while the screen
is off.

So it boosts once successful fingerprint authentication
happened while the screen was off.

To minimize the configuration efforts, this re-uses the strong
boost parameters (boost pattern) from Keypress Boost.

The boost duration can be specified via an overlay and
defaults to 1200ms.

Change-Id: Ibd8d7e20d26c0014cbfe10d8ce7cd42a0a9f7081
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Sagar Shah <sagarshah2326@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>