Fix battery detail view battery icon tint
Now that the detail view is exposed, we see the battery icon in the detail
view which is pulled from the actual tile. The problem here is that Google
made the QS panel white so you have a white icon on a white background. This
keeps everything playing nice and still tints the icon to match the theme.

Change-Id: I26bb0c1d7def7d8d7dfa09181cc9632590daadf1
Signed-off-by: Tab Fitts <tfitts1212@gmail.com>